Write a constant value to a file without buffering (#2874)
* Write a constant value to single file without buffering to remove spikes in memory consumption.

This PR solves an issue of memory consumption reported in #2821
We found that there is a spike in memory consumption when writing a constant into a file (model.constants.bin). This is because all constants into a buffer once and write it to a file at once. This PR changes to write the constant to the file without the buffering. This removes the spike in memory consumption.

---------

Signed-off-by: Haruki Imai <imaihal@jp.ibm.com>
Co-authored-by: Tung D. Le <tung@jp.ibm.com>